requirements.txt
Pillow==12.2.0
PySide6==6.11.0
PyInstaller==6.19.0
qt-material==2.17
